refactor: rename node category 'interface' to 'ui'
Rename the NodeTypeCategory value from 'interface' to 'ui' across
the entire codebase for clarity and consistency. This includes:

- Shared type definition (NodeTypeCategory union)
- Node definitions (BlankComponentNode category, directory rename)
- Backend services and tests (node.service, schema.service)
- Frontend components (AddStepModal, FlowDiagram, NodeLibrary,
  NodeExecutionCard, NodeGroup, NodeItem)
