+++ packages/manifest/backend/src/node/schema/schema.service.ts
@@ -521,7 +521,7 @@ export class SchemaService {
 
       if (targetNode?.type === 'Link') {
         const sourceNodeDef = this.nodeTypeMap.get(sourceNode?.type ?? '');
-        if (sourceNodeDef && sourceNodeDef.category !== 'interface') {
+        if (sourceNodeDef && sourceNodeDef.category !== 'ui') {
           results.push({
             connectionId: connection.id,
             sourceNodeId: connection.sourceNodeId,
@@ -529,7 +529,7 @@ export class SchemaService {
             status: 'error',
             issues: [{
               type: 'type-mismatch',
-              message: `Link nodes can only be connected after UI nodes (interface category). "${sourceNodeDef.displayName}" is a ${sourceNodeDef.category} node.`,
+              message: `Link nodes can only be connected after UI nodes (ui category). "${sourceNodeDef.displayName}" is a ${sourceNodeDef.category} node.`,
               sourcePath: '',
               targetPath: '',
             }],
